create green-invoice module
