+++ packages/server/src/modules/documents/resolvers/documents.resolver.ts
@@ -1,35 +1,16 @@
-import { endOfMonth, format, startOfMonth, subMonths } from 'date-fns';
 import { GraphQLError } from 'graphql';
-import {
-  DocumentInput_Input,
-  Currency as GreenInvoiceCurrency,
-} from '@accounter/green-invoice-graphql';
-import { CloudinaryProvider } from '@modules/app-providers/cloudinary.js';
 import { GoogleDriveProvider } from '@modules/app-providers/google-drive/google-drive.provider.js';
-import { GreenInvoiceProvider } from '@modules/app-providers/green-invoice.js';
 import { deleteCharges } from '@modules/charges/helpers/delete-charges.helper.js';
 import { ChargesProvider } from '@modules/charges/providers/charges.provider.js';
-import { BusinessesGreenInvoiceMatcherProvider } from '@modules/financial-entities/providers/businesses-green-invoice-match.provider.js';
-import { BusinessesProvider } from '@modules/financial-entities/providers/businesses.provider.js';
 import { EMPTY_UUID } from '@shared/constants';
-import { Currency, DocumentType } from '@shared/enums';
+import { DocumentType } from '@shared/enums';
 import { Resolvers } from '@shared/gql-types';
-import {
-  dateToTimelessDateString,
-  formatCurrency,
-  optionalDateToTimelessDateString,
-} from '@shared/helpers';
-import {
-  getGreenInvoiceDocumentType,
-  normalizeDocumentType,
-} from '../helpers/green-invoice.helper.js';
 import { getDocumentFromFile } from '../helpers/upload.helper.js';
 import { DocumentsProvider } from '../providers/documents.provider.js';
 import type {
   DocumentsModule,
   IGetAllDocumentsResult,
   IInsertDocumentsParams,
-  IInsertDocumentsResult,
   IUpdateDocumentParams,
 } from '../types.js';
 import {
@@ -357,240 +338,6 @@ export const documentsResolvers: DocumentsModule.Resolvers &
           }`,
         };
       }
-    },
-    fetchIncomeDocuments: async (
-      _,
-      { ownerId },
-      { injector, adminContext: { defaultAdminBusinessId } },
-    ) => {
-      const data = await injector.get(GreenInvoiceProvider).searchDocuments({
-        input: { pageSize: 100, sort: 'creationDate' },
-      });
-      if (!data?.items) {
-        throw new GraphQLError('Failed to fetch documents');
-      }
-      if (data.items.length === 0) {
-        return [];
-      }
-
-      const documents = await injector.get(DocumentsProvider).getAllDocuments();
-      const newDocuments = data.items.filter(
-        item =>
-          item &&
-          !documents.some(
-            doc =>
-              doc.vat_amount === item.vat &&
-              doc.total_amount === item.amount &&
-              doc.serial_number === item.number &&
-              optionalDateToTimelessDateString(doc.date) === item.documentDate,
-          ) &&
-          item.type !== '_300',
-      );
-      const addedDocs: IInsertDocumentsResult[] = [];
-
-      await Promise.all(
-        newDocuments.map(async greenInvoiceDoc => {
-          if (!greenInvoiceDoc) {
-            return;
-          }
-
-          const documentType = normalizeDocumentType(greenInvoiceDoc.type);
-          const isOwnerCreditor =
-            greenInvoiceDoc.amount > 0 && documentType !== DocumentType.CreditInvoice;
-
-          try {
-            // generate preview image via cloudinary
-            const imagePromise = injector
-              .get(CloudinaryProvider)
-              .uploadInvoiceToCloudinary(greenInvoiceDoc.url.origin);
-
-            // Generate parent charge
-            const chargePromise = injector.get(ChargesProvider).generateCharge({
-              ownerId,
-              userDescription:
-                greenInvoiceDoc.description && greenInvoiceDoc.description !== ''
-                  ? greenInvoiceDoc.description
-                  : 'Green Invoice generated charge',
-            });
-
-            // Get matching business
-            const businessPromise = injector
-              .get(BusinessesGreenInvoiceMatcherProvider)
-              .getBusinessMatchByGreenInvoiceIdLoader.load(greenInvoiceDoc.client.id);
-
-            const [{ imageUrl }, [charge], business] = await Promise.all([
-              imagePromise,
-              chargePromise,
-              businessPromise,
-            ]);
-
-            if (!charge) {
-              throw new Error('Failed to generate charge');
-            }
-
-            const counterpartyId = business?.business_id ?? null;
-
-            // insert document
-            const rawDocument: IInsertDocumentsParams['document']['0'] = {
-              image: imageUrl,
-              file: greenInvoiceDoc.url.origin,
-              documentType,
-              serialNumber: greenInvoiceDoc.number,
-              date: greenInvoiceDoc.documentDate,
-              amount: greenInvoiceDoc.amount,
-              currencyCode: formatCurrency(greenInvoiceDoc.currency),
-              vat: greenInvoiceDoc.vat,
-              chargeId: charge.id,
-              vatReportDateOverride: null,
-              noVatAmount: null,
-              creditorId: isOwnerCreditor ? defaultAdminBusinessId : counterpartyId,
-              debtorId: isOwnerCreditor ? counterpartyId : defaultAdminBusinessId,
-              allocationNumber: null, // TODO: add allocation number from GreenInvoice API
-            };
-
-            const newDocumentPromise = injector
-              .get(DocumentsProvider)
-              .insertDocuments({ document: [rawDocument] });
-
-            const chargeDescriptionUpdate = new Promise(resolve => {
-              const income = greenInvoiceDoc.income;
-              if (
-                !income ||
-                income.length === 0 ||
-                !income[0]?.description ||
-                income[0].description === ''
-              ) {
-                resolve(undefined);
-              }
-
-              const userDescription = income
-                .filter(item => item?.description)
-                .map(item => item!.description)
-                .join(', ');
-
-              injector
-                .get(ChargesProvider)
-                .updateCharge({
-                  chargeId: charge.id,
-                  userDescription,
-                })
-                .then(() => resolve(undefined));
-            });
-
-            const [newDocument] = await Promise.all([newDocumentPromise, chargeDescriptionUpdate]);
-
-            addedDocs.push(newDocument[0]);
-          } catch (e) {
-            throw new GraphQLError(
-              `Error adding Green Invoice document: ${e}\n\n${JSON.stringify(
-                greenInvoiceDoc,
-                null,
-                2,
-              )}`,
-            );
-          }
-        }),
-      );
-
-      return addedDocs;
-    },
-    generateMonthlyClientDocuments: async (_, __, { injector }) => {
-      const jsonArray: Array<{
-        businessId: string;
-        amount: number;
-        currency: GreenInvoiceCurrency;
-      }> = [
-        {
-          businessId: '147d3415-55e3-497f-acba-352dcc37cb8d', // uri test
-          amount: 1000,
-          currency: Currency.Ils,
-        },
-      ];
-
-      const errors: string[] = [];
-
-      const proformaProtos = await Promise.all(
-        jsonArray.map(async json => {
-          const businessPromise = injector
-            .get(BusinessesProvider)
-            .getBusinessByIdLoader.load(json.businessId);
-          const businessGreenInvoiceMatchPromise = injector
-            .get(BusinessesGreenInvoiceMatcherProvider)
-            .getBusinessMatchByIdLoader.load(json.businessId);
-          const [business, businessGreenInvoiceMatch] = await Promise.all([
-            businessPromise,
-            businessGreenInvoiceMatchPromise,
-          ]);
-
-          if (!business) {
-            throw new GraphQLError(`Business ID="${json.businessId}" not found`);
-          }
-
-          if (!businessGreenInvoiceMatch) {
-            throw new GraphQLError(
-              `Green invoice match not found for business ID="${json.businessId}"`,
-            );
-          }
-
-          const today = new Date();
-          const monthStart = dateToTimelessDateString(startOfMonth(today));
-          const monthEnd = dateToTimelessDateString(endOfMonth(today));
-          const year = today.getFullYear() + (today.getMonth() === 0 ? -1 : 0);
-          const month = format(subMonths(today, 1), 'MMMM');
-
-          const documentInput: DocumentInput_Input & { businessName: string } = {
-            businessName: business.name,
-            type: getGreenInvoiceDocumentType(
-              businessGreenInvoiceMatch.document_type as DocumentType,
-            ),
-            remarks: businessGreenInvoiceMatch.remark ?? undefined,
-            date: monthStart,
-            dueDate: monthEnd,
-            lang: 'en',
-            currency: json.currency,
-            vatType: 1,
-            rounding: false,
-            signed: true,
-            attachment: true,
-            client: {
-              id: businessGreenInvoiceMatch.green_invoice_id,
-              emails: [...(businessGreenInvoiceMatch.emails ?? []), 'ap@the-guild.dev'],
-            },
-            income: [
-              {
-                description: `GraphQL Hive Enterprise License - ${month} ${year}`,
-                quantity: 1,
-                price: json.amount,
-                currency: json.currency,
-                vatType: 1,
-              },
-            ],
-          };
-
-          return documentInput;
-        }),
-      );
-
-      for (const proformaProto of proformaProtos) {
-        const { businessName, ...input } = proformaProto;
-        await injector
-          .get(GreenInvoiceProvider)
-          .addDocuments({ input })
-          .then(res => {
-            if (res && 'errorMessage' in res) {
-              errors.push(`${businessName}: ${res.errorMessage}`);
-            }
-          })
-          .catch(e => {
-            console.error(e);
-            errors.push(`${businessName}: ${e.message}`);
-          });
-      }
-
-      return {
-        success: true,
-        errors,
-      };
     },
   },
   Document: {
