+++ packages/server/src/modules/green-invoice/typeDefs/green-invoice.graphql.ts
@@ -1,0 +1,22 @@
+import { gql } from 'graphql-modules';
+
+// eslint-disable-next-line import/no-default-export
+export default gql`
+  extend type Query {
+    greenInvoiceBusinesses: [GreenInvoiceBusiness!]! @auth(role: ACCOUNTANT)
+  }
+  extend type Mutation {
+    fetchIncomeDocuments(ownerId: UUID!): [Document!]! @auth(role: ADMIN)
+    generateMonthlyClientDocuments: GenerateMonthlyClientDocumentsResult! @auth(role: ACCOUNTANT)
+  }
+
+  " business extended with green invoice data "
+  type GreenInvoiceBusiness {
+    id: UUID!
+    originalBusiness: LtdFinancialEntity!
+    greenInvoiceId: UUID!
+    remark: String
+    emails: [String!]!
+    generatedDocumentType: DocumentType!
+  }
+`;
