+++ packages/server/src/modules/green-invoice/types.ts
@@ -1,0 +1,2 @@
+export * from './__generated__/types.js';
+export * from './__generated__/green-invoice.types.js';
